refactor: split classes into different files
